test(ui): give the observer ask flow its own colocated suite
